requirements.txt
fastapi
uvicorn
pypdf
faiss-cpu
sentence-transformers
python-multipart